requirements.txt
Collectfast==0.2.0
Django==1.6.5
Fabric==1.9.1
Pillow==2.4.0
South==0.8.4
Unidecode==0.04.16
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
boto==2.29.1
celery==3.1.13
dj-database-url==0.3.0
dj-static==0.0.5
django-appconf==0.6
django-ckeditor-updated==4.4.0
django-compressor==1.4
django-filter==0.7
django-grappelli==2.5.3
django-jsonfield==0.9.13
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.14
ecdsa==0.11
gunicorn==19.0.0
kombu==3.0.21
paramiko==1.14.0
psycopg2==2.5.3
pycrypto==2.6.1
pystache==0.5.4
python-dateutil==2.2
pytz==2014.4
redis==2.10.1
requests==2.3.0
selenium==2.42.1
six==1.7.3
static==1.0.2
wsgiref==0.1.2
